// src/query/constants/api.ts
export enum ApiEndpoints {
    POSTS = '/posts/',
    SIGNUP = '/auth/signup',
    LOGIN = '/auth/login',
    REFRESH = '/auth/refresh',
    CHECKAUTH = '/auth/check-auth',
    VERIFY = '/auth/verify',
    FORGOTPASSWORD = '/auth/forgot-password',
    VERIFYOTP = '/auth/verify-otp',
    RESETPASSWORD = '/auth/reset-password',
}